+++ src/main/java/com/apicatalog/cborld/encoder/value/TypeValueEncoder.java
@@ -1,6 +1,5 @@
 package com.apicatalog.cborld.encoder.value;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import com.apicatalog.cborld.dictionary.Dictionary;
@@ -16,7 +15,7 @@ public class TypeValueEncoder implements ValueEncoder {
     public DataItem encode(Dictionary dictionary, ValueCursor value, String term, Collection<String> types) {
 
         if (types != null && types.contains(Keywords.TYPE)) {
-            final BigInteger code = dictionary.getCode(value.stringValue());
+            final Integer code = dictionary.getCode(value.stringValue());
     
             if (code != null) {
                 return new UnsignedInteger(code);
