+++ src/main/java/com/apicatalog/cborld/encoder/value/VocabValueEncoder.java
@@ -1,6 +1,5 @@
 package com.apicatalog.cborld.encoder.value;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import com.apicatalog.cborld.dictionary.Dictionary;
@@ -16,9 +15,9 @@ public class VocabValueEncoder implements ValueEncoder {
     public DataItem encode(Dictionary dictionary, ValueCursor value, String term, Collection<String> types) {
 
         if (value.isString() && types != null && types.contains(Keywords.VOCAB)) {
-            
-            BigInteger code = dictionary.getCode(value.stringValue());
-            
+
+            Integer code = dictionary.getCode(value.stringValue());
+
             if (code != null) {
                 return new UnsignedInteger(code);
             }
